chore: split the calculations into more managable files
